# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API for hospital management system",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list all appointements",
                "description": "List Appointements Endpoint.",
                "operationId": "listAppointements",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/doctors/{id}/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list doctor's appointements",
                "description": "Doctor's Appointements Endpoint.",
                "operationId": "listDoctorAppointements",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "doctor's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Object Not Found"
                    },
                    "403": {
                        "description": "Not Authorized"
                    }
                }
            }
        },
        "/api/appointements/patients/{id}/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list patient's appointements",
                "description": "Patient's Appointements Endpoint.",
                "operationId": "readPatientAppointements",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "patient's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/appointements/clinics/{id}/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list clinic's appointments",
                "description": "Clinic's Appointement Endpoint.",
                "operationId": "listClinicAppointements",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "clinic's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/appointements/me": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "list current user's  appointements",
                "description": "Current Appointements Endpoint.",
                "operationId": "listCurrentAppointements",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/me/{id}": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "read current user's  appointement",
                "description": "Read  Current Appointement Endpoint.",
                "operationId": "readCurrentAppointements",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "appointement's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Object Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/patients": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "list patients appointements",
                "description": "Patients Appointement Endpoint.",
                "operationId": "listPatientAppointements",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/me/patients/{id}": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "read patient's appointement",
                "description": "Read Patient Appointement Endpoint.",
                "operationId": "readPatientAppointement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "appointement's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Object Not Found"
                    },
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/": {
            "post": {
                "tags": [
                    "Patient"
                ],
                "summary": "create appointement",
                "description": "Create Appointement Endpoint.",
                "operationId": "createAppointement",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clinic_id",
                                    "date"
                                ],
                                "properties": {
                                    "clinic_id": {
                                        "type": "integer"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/appointements/schedule/clinics/{id}": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "schedule clinic appointements",
                "description": "Clinic Schedule Endpoint.",
                "operationId": "clinicSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "clinic's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "appointement's period",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Object Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/schedule/doctors/{id}": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "schedule doctor appointements",
                "description": "Doctor Schedule Endpoint.",
                "operationId": "doctorSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "doctor's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "appointement's period",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Object Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/appointements/me/patients/{id}/submit": {
            "put": {
                "tags": [
                    "Doctor"
                ],
                "summary": "submit appointement's status",
                "description": "Submit Appointement Endpoint.",
                "operationId": "submitSchedule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "patient's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "next_date",
                                    "status"
                                ],
                                "properties": {
                                    "next_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}